requirements.txt
google-adk==1.18.0
python-dotenv
toolbox-core 
langchain
google-genai
a2a-sdk
google-cloud-aiplatform[adk,agent_engines]==1.126.1
locust
google-cloud-spanner==3.56.0
google-adk[eval]
python-a2a
gradio
google-cloud-bigquery
aiohttp
google-cloud-storage
google-cloud-alloydb-connector[pg8000] 
sqlalchemy
PyJWT[crypto]
fastmcp==2.12.5
flask==3.0.0
gunicorn==21.2.0
pypdf